Traktor: Fixed PS3 renderer. Also fixed a lot of static analysiz issues found by SN compiler. Improved PS3 deployment scripts, able to reduce size of ELF significantly.
[git-p4: depot-paths = "//depot/traktor/": change = 19565]

+++ code/Drawing/Raster.cpp
@@ -203,7 +203,7 @@ void Raster::drawPolygon(const Vector2* points, uint32_t npoints, const Color4f&
 			continue;
 		}
 
-		int32_t fillDelta = 1;
+		int8_t fillDelta = 1;
 		if (y1 > y2)
 		{
 			std::swap(x1, x2);
